+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 0.14-97
 - Added fuel gauge for the buzzsaw (BluSunrize)
+- Added a covered version of the splitting conveyor (BluSunrize)
 - Fixed a crash triggered by other mod's automatic recipe unlocks (BluSunrize)
 - Fixed a crash with the wirecoil tooltips (Malte)
 - Fixed a crash when rendering coresamples with Optifine installed (Malte)